refactor: 診断機能を削除し川柳ゲームに特化 🎯
## 変更内容

### 診断機能の完全削除とアーカイブ
- すべての診断関連コード（AI診断、Prairie Card解析）を `/archive/diagnosis-feature/` に移動
- Prairie Card関連のすべての参照と機能を削除
- 診断結果表示、共有機能、QRスキャン機能をアーカイブ

### 川柳ゲーム専用アプリへのリブランディング
- ランディングページを川柳ゲーム専用に完全刷新
- 3つのゲームモード（ソロプレイ、対戦、ランキング）を明確に表示
- アプリ説明を「川柳ゲーム」として統一

### ドキュメントの更新
- README.mdを川柳ゲーム専用に書き換え
- CLAUDE.mdから診断関連の記述を削除
- package.json の description とキーワードを更新

### 依存関係のクリーンアップ
- 不要なパッケージを削除:
  - openai（AI診断用）
  - cheerio（HTML解析用）
  - dompurify（HTMLサニタイゼーション用）
  - qr-scanner（Prairie Card QRスキャン用）
- ※ qrcode は川柳ゲームのルーム共有機能で使用するため保持

### 環境設定の簡素化
- .env.example から診断関連の環境変数を削除
- wrangler.toml から DIAGNOSIS_KV バインディングを削除
- tsconfig.json に archive ディレクトリを除外設定

### アーカイブされた主要ファイル
- /src/app/duo/（2人診断ページ）
- /src/app/group/（グループ診断ページ）
- /src/components/diagnosis/（診断関連コンポーネント）
- /src/lib/diagnosis-*（診断エンジン群）
- /src/lib/prairie-*（Prairie Card関連）
- /functions/api/diagnosis.js, prairie.js（API関数）

## 影響範囲
- ビルド成功確認済み ✅
- 診断機能を完全に削除、川柳ゲーム機能のみ残存
- Git履歴はアーカイブディレクトリに保存
