+++ .env.example
@@ -1,15 +1,7 @@
-# CND² アプリケーション環境設定
-
-# OpenAI API設定（必須）
-OPENAI_API_KEY=your_openai_api_key_here
+# CND² 川柳ゲームアプリケーション環境設定
 
 # アプリケーションURL
 NEXT_PUBLIC_APP_URL=https://cnd2.cloudnativedays.jp
-
-# 診断エンジンバージョン（オプション）
-# v3エンジンを無効にする場合のみ'false'を設定
-# デフォルト: v3エンジンを使用（より面白い診断結果、ラッキーアイテム付き）
-# NEXT_PUBLIC_USE_DIAGNOSIS_V3=false
 
 # Sentry設定（オプション）
 # SENTRY_DSN=your_sentry_dsn
@@ -17,3 +9,6 @@ NEXT_PUBLIC_APP_URL=https://cnd2.cloudnativedays.jp
 
 # 開発環境設定（オプション）
 # NODE_ENV=development
+
+# 川柳管理API認証トークン（管理画面用、オプション）
+# SENRYU_ADMIN_TOKEN=your_admin_token_here
