+++ archive/diagnosis-feature/functions/api/diagnosis-v4-openai.js
