+++ archive/diagnosis-feature/src/app/api/route.ts
