+++ archive/diagnosis-feature/src/app/duo/loading.tsx
