+++ archive/diagnosis-feature/src/app/group/loading.tsx
