+++ archive/diagnosis-feature/src/components/diagnosis/AstrologicalDiagnosisResult.tsx
