+++ archive/diagnosis-feature/src/components/diagnosis/DiagnosisFullDebug.tsx
