+++ archive/diagnosis-feature/src/components/diagnosis/DiagnosisLoadingAnimation.tsx
