+++ archive/diagnosis-feature/src/components/diagnosis/DiagnosisResult.tsx
