+++ archive/diagnosis-feature/src/components/diagnosis/__tests__/DiagnosisFullDebug.test.tsx
