+++ archive/diagnosis-feature/src/components/diagnosis/__tests__/DiagnosisResult.test.tsx
