+++ archive/diagnosis-feature/src/hooks/__mocks__/useQRScannerV2.ts
