+++ archive/diagnosis-feature/src/hooks/useQRScannerV2.ts
