+++ archive/diagnosis-feature/src/lib/diagnosis-engine-v4-openai.ts
