+++ archive/diagnosis-feature/src/lib/prompts/diagnosis-prompts.ts
