+++ archive/diagnosis-feature/src/lib/sanitizer.ts
