+++ archive/diagnosis-feature/src/types/diagnosis-metadata.ts
