+++ archive/diagnosis-feature/src/types/diagnosis.ts
