+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "cnd2-app",
   "version": "1.12.1",
-  "description": "CND² - CloudNative Days × Connect 'n' Discover - エンジニアの出会いを、データで可視化する相性診断アプリ",
+  "description": "CND² - CloudNative Days Winter 2025 公式川柳ゲームアプリ",
   "author": "つかまん (@tsukaman)",
   "license": "Apache-2.0",
   "private": true,
@@ -12,9 +12,8 @@
   "keywords": [
     "cloudnative",
     "cndw2025",
-    "prairie-card",
-    "compatibility",
-    "diagnosis",
+    "senryu",
+    "game",
     "nextjs",
     "typescript"
   ],
@@ -41,16 +40,12 @@
     "@react-three/drei": "^10.7.6",
     "@react-three/fiber": "^9.3.0",
     "@sentry/nextjs": "^10.15.0",
-    "cheerio": "^1.1.2",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "dompurify": "^3.2.7",
     "framer-motion": "^12.23.22",
     "lucide-react": "^0.544.0",
     "nanoid": "^5.1.6",
     "next": "15.5.4",
-    "openai": "^5.23.0",
-    "qr-scanner": "^1.4.2",
     "qrcode": "^1.5.4",
     "react": "19.1.1",
     "react-confetti": "^6.4.0",
